+++ LightNing/src/LN/Renderer/Buffer.h
@@ -81,7 +81,7 @@ namespace LN
 		virtual void SetLayout(const VertexBufferLayout& layout) = 0;
 		virtual const VertexBufferLayout& GetBufferLayout() = 0;
 
-		static Ref<VertexBuffer> Create(float* vertex, size_t size);
+		static Ref<VertexBuffer> Create(float* vertex, uint32_t size);
 	};
 
 
@@ -95,7 +95,7 @@ namespace LN
 
 		virtual int GetCount()const = 0;
 
-		static Ref<IndexBuffer> Create(size_t* index, size_t size);
+		static Ref<IndexBuffer> Create(uint32_t* index, uint32_t size);
 	};
 
 }
